refactor: Suggestion component using typescript
